acu78070 - canceling sleep timers for timeout errors
  Timeout.timeout(1) { # timeout t1
     sleep(10)
  }
After one second, EM schedules the timeout timer
for t1 which calls Fiber.resume(Timeout::Error). 
This raises an exception at the place where we
yielded which is in reactor.rb:

  def self.sleep(time)
    NB::Fiber.yield if time.nil?
    return if time <= 0
    fiber = NB::Fiber.current
    NB.reactor.add_timer(time){ fiber.resume} # *** sleep timer ***
    NB::Fiber.yield # **** exception gets raised here ****
  end

unlike for timeout timers, we don't cancel this 
sleep timer with the timeout error. As a result, 
when this timer hits, fiber.resume will be called 
which will cause the fiber to resume even though 
the fiber might have already moved on to a different
task and might be waiting on a different yield
which should not be called yet. Example:

begin
  Timeout.timeout(1) { # timeout t1
     sleep(10)
  }
rescue => e
  puts "e.message: #{e.message}"
end

Timeout.timeout(1) { # timeout t2
   sleep(500) 
}

in this example the sleep timer from the t1 timer 
block will still be pending after catching the 
TimeoutError. Even though a TimeoutError was raised
t1's sleep timer will still fire after 9 seconds 
and will call resume. Unfortunately, in the meanwhile
a new sleep has been scheduled and the fiber has
yielded. As a result, the sleep(500)'s yield will 
get resumed after 9 seconds instead of 500 seconds.