acu78070 - added failing spec to show problem with EM.many_ticks not cleaning up its state between runs.
If a tick is scheduled with EM.many_ticks and EM is stopped and started again. EM.many_ticks' state is not cleaned up and when EM is restarted. The scheduled EM.many_ticks tick wont ever be run because EM has been stopped. Any new ticks scheduled with EM.many_ticks wont ever run because EM.many_ticks thinks that a tick is still pending.