acu85369 - removing & detaching the correct handler on unbind
When detach is called on a handler, unbind gets called
asyncly.
If a new handler gets created for the same fd, then
the unbind can detach the newly created handler.

Instead of detaching whatever handler is currently
mapped to the fd, we check if the handler matches
before removing the handler and detaching it.